Restructure shared components i.e., Info and Privacy Policy; rebuild Privacy Policy content.
